requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
gluonts==0.14.4
google-pasta==0.2.0
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
inquirerpy==0.3.4
Jinja2==3.1.5
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
lag-llama @ git+https://github.com/time-series-foundation-models/lag-llama.git@37cce619728f36116fa06efe1b05765655b7e5c7
libclang==18.1.1
lightning==2.1.4
lightning-utilities==0.11.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
networkx==3.4.2
nixtla==0.6.5
numpy==1.26.4
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.14
packaging==24.2
pandas==2.1.4
patsy==1.0.1
pfzy==0.3.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pydantic==2.10.5
pydantic_core==2.27.2
pyEsn==0.0.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.1.4
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentry-sdk==2.20.0
setproctitle==1.3.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toolz==0.12.1
torch==2.5.1
torchmetrics==1.6.1
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
utilsforecast==0.2.10
wandb==0.19.4
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
